move workpool to own file and configure default retries
