+++ src/component/_generated/api.d.ts
@@ -12,6 +12,7 @@ import type * as functions from "../functions.js";
 import type * as journal from "../journal.js";
 import type * as logging from "../logging.js";
 import type * as model from "../model.js";
+import type * as pool from "../pool.js";
 import type * as sleep from "../sleep.js";
 import type * as utils from "../utils.js";
 import type * as workflow from "../workflow.js";
@@ -35,6 +36,7 @@ declare const fullApi: ApiFromModules<{
   journal: typeof journal;
   logging: typeof logging;
   model: typeof model;
+  pool: typeof pool;
   sleep: typeof sleep;
   utils: typeof utils;
   workflow: typeof workflow;
